Add combined, merged, and replaced.
Combine documents. Combined is a curried function. It's first parameter is a function that resolves overlapping nodes. merged and replaced uses it under the hood.

merged combines two documents and resolves overlapping nodes into a simultainity.

replaced combines two documents and resolves overlapping by replacing the value of the target document with the overlapping document.